requirements.txt
dotenv
matplotlib
pillow
Flask
azure-ai-vision-imageanalysis==1.0.0
requests
uuid